Get rid of the form_as_fieldsets.html template.
