--- uswds_forms/templates/uswds_forms/form_as_fieldsets.html
@@ -1,4 +1,0 @@
-{% load uswds_forms %}
-{% for field in form %}
-  {% fieldset field %}
-{% endfor %}
