+++ uswds_forms/tests/test_jinja2.py
@@ -24,3 +24,13 @@ def test_input_option_template_is_identical():
 
     assert (JINJA2_DIR / 'input_option.html').read_text() == \
            (TEMPLATES_DIR / 'input_option.html').read_text()
+
+
+def test_fieldset_template_is_identical():
+    '''
+    This template uses the subset of syntax shared by Jinja2 and
+    Django Templates, so the two versions of it should be identical.
+    '''
+
+    assert (JINJA2_DIR / 'fieldset.html').read_text() == \
+           (TEMPLATES_DIR / 'fieldset.html').read_text()
